admin category crud

+++ app/Http/Controllers/Admin/ProductController.php
@@ -136,8 +136,10 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
-        Storage::disk('local')->delete($product->picture->link);
-        $product->picture()->delete();
+        if ($product->picture) {
+            Storage::disk('local')->delete($product->picture->link);
+            $product->picture()->delete();
+        }
 
         Product::destroy($product->id);
